# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GIAM (General Identity and Access Manager)",
    "version": "1.0.0",
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "clientBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "mtls": {
        "type": "mutualTLS"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "description": "RFC 9457 problem details.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI identifying the problem type.",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary.",
            "example": "Not Found"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code.",
            "example": 404
          },
          "detail": {
            "type": "string",
            "description": "An explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "description": "A URI identifying this occurrence."
          }
        },
        "title": "Problem",
        "example": {
          "type": "about:blank",
          "title": "Not Found",
          "status": 404
        }
      },
...
    }
  },
  "paths": {
...
    "/realms/{realm}/grants": {
      "get": {
        "operationId": "listGrants",
        "summary": "What this principal has authorised",
        "tags": [
          "consent"
        ],
        "description": "Standard-adjacent: the grants behind RFC 6749 authorisation, exposed so a person can review and withdraw them. A revoked grant is kept rather than deleted, because \"what did I once allow, and when did I stop\" is the question that matters after something goes wrong.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "revoked",
                "all"
              ],
              "default": "all"
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "subjectId",
            "required": false,
            "description": "Another principal, for a caller whose role grants oversight. Refused otherwise."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "accountHolderRef",
            "required": false,
            "description": "The principal bound to this business reference. Same oversight rule as subjectId."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "clientId",
            "required": false,
            "description": "Everyone who has authorised this client. An oversight query, so it needs the same permission."
          },
          {
            "schema": {
              "type": "string"
            },
            "example": "acme",
            "in": "path",
            "name": "realm",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authorisations held.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The authorisations held.",
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "grants"
                  ],
                  "properties": {
                    "grants": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "grantId",
                          "clientId",
                          "clientName",
                          "scopes",
                          "status",
                          "grantedAt"
                        ],
                        "properties": {
                          "grantId": {
                            "type": "string"
                          },
                          "clientId": {
                            "type": "string"
                          },
                          "clientName": {
                            "type": "string"
                          },
                          "logoUri": {
                            "type": "string"
                          },
                          "scopes": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "revoked"
                            ]
                          },
                          "grantedAt": {
                            "type": "string"
                          },
                          "revokedAt": {
                            "type": "string"
                          },
                          "lastUsedAt": {
                            "type": "string"
                          }
                        },
                        "example": {
                          "grantId": "g-4c1f",
                          "clientId": "acme-portal",
                          "clientName": "Acme Portal",
                          "scopes": [
                            "openid",
                            "profile"
                          ],
                          "status": "active",
                          "grantedAt": "2026-08-01T10:22:00.000Z"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "grants": [
                    {
                      "grantId": "g-4c1f",
                      "clientId": "acme-portal",
                      "clientName": "Acme Portal",
                      "scopes": [
                        "openid",
                        "profile"
                      ],
                      "status": "active",
                      "grantedAt": "2026-08-01T10:22:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No valid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "No role held grants this oversight.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "No such realm.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
    "/realms/{realm}/grants/{grantId}/operations": {
      "get": {
        "operationId": "listGrantOperations",
        "summary": "What was done under this authorisation",
        "tags": [
          "consent"
        ],
        "description": "No applicable standard. The identity trail for one grant, read from the security event record. What a client DID with the access is the consuming application's business trail and stays there: recording it here would create a second source of truth for it.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "subjectId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "accountHolderRef",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "default": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "realm",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "grantId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The identity events recorded under this authorisation.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The identity events recorded under this authorisation.",
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "operations"
                  ],
                  "properties": {
                    "operations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "ts": {
                            "type": "string"
                          },
                          "action": {
                            "type": "string"
                          },
                          "outcome": {
                            "type": "string"
                          },
                          "clientId": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "operations": [
                    {
                      "ts": "2026-08-20T08:04:11.000Z",
                      "action": "token.issued",
                      "outcome": "success",
                      "clientId": "acme-portal"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No valid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "No role held grants this oversight.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "No such grant for this principal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/realms/{realm}/grants/{grantId}/reactivate": {
      "post": {
        "operationId": "reactivateGrant",
        "summary": "Restore a withdrawn authorisation",
        "tags": [
          "consent"
        ],
        "description": "Standard-adjacent: withdrawal of the consent behind RFC 6749 authorisation. Restores a previously withdrawn authorisation without a fresh approval round, which is why it is the owner's own action and nobody else's.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "realm",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "grantId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authorisation, in its new state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "grantId",
                    "clientId",
                    "clientName",
                    "scopes",
                    "status",
                    "grantedAt"
                  ],
                  "properties": {
                    "grantId": {
                      "type": "string"
                    },
                    "clientId": {
                      "type": "string"
                    },
                    "clientName": {
                      "type": "string"
                    },
                    "logoUri": {
                      "type": "string"
                    },
                    "scopes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "revoked"
                      ]
                    },
                    "grantedAt": {
                      "type": "string"
                    },
                    "revokedAt": {
                      "type": "string"
                    },
                    "lastUsedAt": {
                      "type": "string"
                    }
                  },
                  "description": "The authorisation, in its new state."
                },
                "example": {
                  "grantId": "g-4c1f",
                  "clientId": "acme-portal",
                  "clientName": "Acme Portal",
                  "scopes": [
                    "openid",
                    "profile"
                  ],
                  "status": "active",
                  "grantedAt": "2026-08-01T10:22:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "No role held permits acting for another principal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "No withdrawn grant with that id for this principal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8085",
      "description": "private, service to service"
    }
  ],
...
}